file path choose option added and delete file issue

+++ src/index.js
@@ -5,6 +5,17 @@ module.exports = {
     const { PixelbinConfig, PixelbinClient } = require("@pixelbin/admin");
     const pixelbinConfig = new PixelbinConfig(config);
     const pixelbinClient = new PixelbinClient(pixelbinConfig);
+    const DEFAULT_PATH = "strapi-images";
+    const folderPath = config.folderName || DEFAULT_PATH;
+    
+    // Helper function to extract the data after 'original/'
+    const extractData = (url) => {
+      const originalIndex = url.indexOf("original/");
+      if (originalIndex !== -1) {
+        return url.substring(originalIndex + "original/".length);
+      }
+      return null;
+    };
 
     return {
       async upload(file, customParams = {}) {
@@ -25,6 +36,8 @@ module.exports = {
           options: {
             originalFilename: file.name,
           },
+          path: folderPath,
+          name: file.name,
           overwrite: true,
           ...customParams,
         };
@@ -42,18 +55,24 @@ module.exports = {
       },
 
       async delete(file, customParams = {}) {
-        const requiredProperties = ["hash", "ext"];
+        const requiredProperties = ["url"];
         if (
           !file ||
           !requiredProperties.every((prop) => file[prop] !== undefined)
         ) {
           throw new Error(
-            "Invalid file object. Make sure it has hash and ext properties for delete operation."
+            "Invalid file object. Make sure it has url property for delete operation."
           );
         }
 
+        // Extract the relevant part of the URL
+        const fileId = extractData(file.url);
+        if (!fileId) {
+          throw new Error("Invalid file URL. Unable to extract file ID.");
+        }
+
         const response = await pixelbinClient.assets.deleteFile({
-          fileId: `${file.hash}.${file.ext}`,
+          fileId: `${fileId}`,
           ...customParams,
         });
